Finished FEN parsing + minor refactoring/cleanup

+++ engine/include/Engine.hpp
@@ -1,3 +1,5 @@
+//TODO: comments
+
 #pragma once
 
 #include "board/Board.hpp"
@@ -6,7 +8,7 @@ class Engine {
 private:
     Board* board;
     bool whiteTurn = true;
-    short drawMoveCounter = 0;
+    int drawMoveCounter = 0;
 
 public:
     //constructors/destructor
